+++ src/main/java/com/jcg/springmvc/mongo/model/User.java
@@ -1,0 +1,61 @@
+package com.jcg.springmvc.mongo.model;
+
+import java.io.Serializable;
+
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document(collection = "users")
+public class User implements Serializable {
+
+	private static final long serialVersionUID = 1L;
+
+	private String id;
+	
+	private String name, email, password;
+
+	public User() {
+		super();
+	}
+
+	public User(String id, String name, String email, String password) {
+		super();
+		this.id = id;
+		this.name = name;
+		this.email = email;
+		this.password = password;
+	}
+
+	public String getId() {
+		return id;
+	}
+
+	public void setId(String id) {
+		this.id = id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+	
+}
