+++ src/main/webapp/WEB-INF/views/registration.jsp
@@ -26,19 +26,19 @@
 
 					<div class="wrap-input100 validate-input m-b-23" data-validate = "Username is reauired">
 						<span class="label-input100">Username</span>
-						<input class="input100" type="text" path="name" id="name" placeholder="Type your username">
+						<input class="input100" type="text" path="name" id="name" name="name"  placeholder="Type your username">
 						<span class="focus-input100" data-symbol="&#xf206;"></span>
 					</div>
 					
 					<div class="wrap-input100 validate-input" data-validate="Password is required">
 						<span class="label-input100">Email</span>
-						<input class="input100" type="text" path="email" id="email" placeholder="Type your email">
+						<input class="input100" type="text" path="email" id="email" name="email"  placeholder="Type your email">
 						<span class="focus-input100" data-symbol="&#xf190;"></span>
 					</div>
 
 					<div class="wrap-input100 validate-input" data-validate="Password is required">
 						<span class="label-input100">Password</span>
-						<input class="input100" type="password" path="password" id="password" placeholder="Type your password">
+						<input class="input100" type="password" path="password" id="password" name="password" placeholder="Type your password">
 						<span class="focus-input100" data-symbol="&#xf190;"></span>
 					</div>
